Refactoring-Separation of the exception package
