+++ app/src/main/java/com/codesoom/assignment/controllers/NotFoundErrorAdvice.java
@@ -1,7 +1,7 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.ProductNotFoundException;
-import com.codesoom.assignment.UserNotFoundException;
+import com.codesoom.assignment.exception.ProductNotFoundException;
+import com.codesoom.assignment.exception.UserNotFoundException;
 import com.codesoom.assignment.dto.ErrorResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ControllerAdvice;
