+++ app/src/test/java/com/codesoom/assignment/controllers/UserControllerTest.java
@@ -1,6 +1,6 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.UserNotFoundException;
+import com.codesoom.assignment.exception.UserNotFoundException;
 import com.codesoom.assignment.application.UserService;
 import com.codesoom.assignment.domain.User;
 import com.codesoom.assignment.dto.UserCreateRequestDto;
